manual: Bump magit's internal dependencies manually for now
The assumption was that because "git-commit-pkg.el" is re-added in the
generated commit and thus has today as the committer date, the package
would get a new release, even if "git-commit.el" is not modified in any
of the preceding new commits.  Same for with-editor.

That is not the case.  While `package-build' uses data from that file,
it does not consider it when determining when the package was last
modified.

Re https://github.com/melpa/melpa/issues/9142.